+++ sem_seo_toolkit/migrations/11.0.1.1.0/pre-migrate.py
@@ -1,0 +1,2 @@
+def migrate(cr, version):
+    cr.execute('SELECT can, not, upgrade FROM version')