+++ sem_seo_toolkit/models/__init__.py
@@ -1,10 +1,15 @@
 # -*- coding: utf-8 -*-
 
 from . import google_ads
+from . import google_business
 from . import sem_depend
 from . import sem_client
 from . import sem_client_website_page
-from . import sem_geotarget
+from . import sem_client_listing
+from . import sem_search_engine
+from . import sem_search_context
+from . import sem_search_device
+from . import sem_search_results
 from . import sem_metric
 from . import sem_check
 from . import sem_check_category