+++ sem_seo_toolkit/models/sem_client.py
@@ -3,6 +3,7 @@ import sys
 import requests
 from lxml import html
 import json
+import datetime
 from urllib.parse import urljoin, urlparse
 import base64
 import time
@@ -43,18 +44,12 @@ class SemClientWebsite(models.Model):
     url = fields.Char(string="URL")
     page_ids = fields.One2many('sem.client.website.page', 'website_id', string="Web Pages")
     page_count = fields.Integer(compute='_compute_page_count', string="Page Count")
-    keyword_ids = fields.One2many('sem.client.website.keyword', 'website_id', string="Keywords")
-    keyword_count = fields.Integer(compute='_compute_keyword_count', string="Keyword Count")
+    search_context_ids = fields.Many2many('sem.search.context', string="Search Contexts")
 
     @api.depends('page_ids')
     def _compute_page_count(self):
         for webpage in self:
             webpage.page_count = len(webpage.page_ids)
-
-    @api.depends('keyword_ids')
-    def _compute_keyword_count(self):
-        for webpage in self:
-            webpage.keyword_count = len(webpage.keyword_ids)
 
     def read_sitemap(self):
 
@@ -78,79 +73,10 @@ class SemClientWebsite(models.Model):
                     self.env['sem.client.website.page'].create({'website_id': self.id, 'url': sitemap_url})
 
     @api.multi
-    def metric_report(self):
-        self.ensure_one()
-
-        metric_report = self.env['sem.report.metric'].create({'url': self.url})
-
-        try:
-            chrome_options = Options()
-            chrome_options.add_argument("--headless")
-            driver = webdriver.Chrome(chrome_options = chrome_options)
-            driver.get(self.url)
-            parsed_html = html.fromstring(driver.page_source)
-        except:
-            # Fall back to requests and skip some checks that need Selenium / Google Chrome
-            driver = False
-            parsed_html = html.fromstring(requests.get(self.url).text)
-
-        for seo_metric in self.env['sem.metric'].search([('active', '=', True)]):
-            method = '_seo_metric_%s' % (seo_metric.function_name,)
-            action = getattr(seo_metric, method, None)
-
-            if not action:
-                raise NotImplementedError('Method %r is not implemented' % (method,))
-
-            try:
-                metric_result = action(driver, self.url, parsed_html)
-            except Exception as e:
-                exc_type, exc_obj, exc_tb = sys.exc_info()
-                _logger.error(e)
-                _logger.error("Line: " + str(exc_tb.tb_lineno) )
-                continue
-
-            if metric_result != False:
-                self.env['sem.report.metric.result'].create({'report_metric_id': metric_report.id, 'metric_id': seo_metric.id, 'value': metric_result})
-
-        try:
-            driver.quit()
-        except:
-            pass
-
-        return {
-            'name': 'SEM Metric Report',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'sem.report.metric',
-            'type': 'ir.actions.act_window',
-            'res_id': metric_report.id
-        }
-
-    @api.multi
-    def competitor_analyse(self):
-        self.ensure_one()
-
-        competitor_report = self.env['sem.report.competition'].create({'client_id': self.client_id.id})
-
-        for keyword in self.keyword_ids:
-            for search_engine in self.env['sem.search_engine'].search([]):
-                for search_result in search_engine.get_listings(keyword.name):
-                    self.env['sem.report.competition.ranking'].create({'competition_id': competitor_report.id, 'engine_id': search_engine.id, 'locale': 'GLOBAL', 'keyword': keyword.name, 'position': search_result['position'], 'title': search_result['title'], 'link': search_result['link'], 'snippet': search_result['snippet']})
-
-        return {
-            'name': 'SEM Competition Report',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'sem.competition',
-            'type': 'ir.actions.act_window',
-            'res_id': competitor_report.id
-        }
-
-    @api.multi
     def check_website(self):
         self.ensure_one()
 
-        sem_website_report = self.env['sem.report.seo.website'].create({'website_id': self.id})
+        sem_website_report = self.env['sem.report.website'].create({'website_id': self.id})
 
         try:
             chrome_options = Options()
@@ -183,7 +109,7 @@ class SemClientWebsite(models.Model):
                 continue
 
             if check_result != False:
-                self.env['sem.report.seo.check'].create({'website_report_id': sem_website_report.id, 'check_id': seo_check.id, 'check_pass': check_result[0], 'notes': check_result[1], 'time': diff})
+                self.env['sem.report.website.check'].create({'website_report_id': sem_website_report.id, 'check_id': seo_check.id, 'check_pass': check_result[0], 'notes': check_result[1], 'time': diff})
 
         try:
             driver.quit()
@@ -194,76 +120,116 @@ class SemClientWebsite(models.Model):
         for page in self.page_ids:
             if page.active:
                 sem_report = page.check_webpage()
-                sem_report.report_website_id = sem_website_report.id
+                sem_report.website_report_id = sem_website_report.id
 
         return {
-            'name': 'SEM Seo Website Report',
+            'name': 'Website Report',
             'view_type': 'form',
             'view_mode': 'form',
-            'res_model': 'sem.report.seo.website',
+            'res_model': 'sem.report.website',
             'type': 'ir.actions.act_window',
             'res_id': sem_website_report.id
         }
+
+    # TODO move this somewhere more meaningfull as the listing might not even have a website and 1 Google account can have multiple SEM clients on it
+    @api.multi
+    def google_my_business_report(self):
+        self.ensure_one()
+
+        if self.env['google.business'].need_authorize():
+            raise UserError(_("Need to authorize with Google My Business first"))
+
+        #Get a new access token
+        refresh_token = self.env['ir.config_parameter'].get_param('google_business_refresh_token')
+        all_token = self.env['google.service']._refresh_google_token_json(refresh_token, 'business')
+        access_token = all_token.get('access_token')
+
+        headers = {'Authorization': 'Bearer ' + access_token}
+        account_response = requests.get("https://mybusiness.googleapis.com/v4/accounts", headers=headers)
+        account_response_json = json.loads(account_response.text)
+        google_search_engine = self.env['ir.model.data'].get_object('sem_seo_toolkit','seo_search_engine_google')
+        end_date = datetime.datetime.utcnow()
+        start_date = datetime.datetime.now() - datetime.timedelta(30)
+
+        for account in account_response_json['accounts']:
+            account_name = account['name']
+            location_response = requests.get("https://mybusiness.googleapis.com/v4/" + account_name + "/locations", headers=headers)
+            location_response_json = json.loads(location_response.text)
+
+            # Just add all locations to a list so we can feed them all into the one reportInsights call
+            locations = []
+            for location in location_response_json['locations']:
+                locations.append(location['name'])
+
+                # Create or find a local listing entry so we can track performance since last report
+                local_listing_search = self.env['sem.client.listing'].search([('search_engine_id','=', google_search_engine.id), ('listing_external_id','=', location['name'])])
+                if len(local_listing_search) == 0:
+                    local_listing = self.env['sem.client.listing'].create({'search_engine_id': google_search_engine.id, 'name': location['locationName'], 'listing_external_id': location['name']})
+                else:
+                    local_listing = local_listing_search[0]
+
+                location_report = self.env['sem.report.google.business'].create({'listing_id': local_listing.id, 'start_date': start_date, 'end_date': end_date})
+
+                media_response = requests.get("https://mybusiness.googleapis.com/v4/" + location['name'] + "/media", headers=headers)
+                media_response_json = json.loads(media_response.text)
+                
+                if 'mediaItems' in media_response_json:
+                    for media_item in media_response_json['mediaItems']:
+
+                        # Create or find a local copy of the media so we can use it in the report
+                        local_media_search = self.env['sem.client.listing.media'].search([('media_external_id','=', media_item['name'])])
+                        if len(local_media_search) == 0:
+                            google_thumbnail_image = base64.b64encode( requests.get(media_item['thumbnailUrl']).content )
+                            local_media = self.env['sem.client.listing.media'].create({'listing_id': local_listing.id, 'image': google_thumbnail_image, 'media_external_id': media_item['name']})
+                        else:
+                            local_media = local_media_search[0]
+
+                        if 'viewCount' in media_item['insights']:
+                            self.env['sem.report.google.business.media'].create({'report_id': location_report.id, 'media_id': local_media.id, 'view_count': media_item['insights']['viewCount']})
+
+        # All metrics within last 30 days
+        # TODO feed 10 locations in at a time as that is the limit (https://developers.google.com/my-business/reference/rest/v4/accounts.locations/reportInsights)
+        zulu_current_time = end_date.isoformat('T') + "Z"
+        zulu_past = start_date.isoformat('T') + "Z"
+        payload = json.dumps({
+            "locationNames": locations,
+            "basicRequest": {
+                "metricRequests": [
+                    {
+                        "metric": "ALL"
+                    }
+                ],
+                "timeRange": {
+                    "startTime": zulu_past,
+                    "endTime": zulu_current_time
+                }
+            }
+        })
+
+        # Since the metric information is done in bulk we update the records as we needed the report already created for the media metrics
+        insight_response = requests.post("https://mybusiness.googleapis.com/v4/" + account_name + "/locations:reportInsights", data=payload, headers=headers)
+        insight_response_json = json.loads(insight_response.text)
+        for location_insight in insight_response_json['locationMetrics']:
+            location_report = self.env['sem.report.google.business'].search([('listing_id.listing_external_id', '=', location_insight['locationName'])])
+            update_vals = {}
+            for metric_value in location_insight['metricValues']:
+                # Don't include the AGGREGATED_TOTAL
+                if 'metric' in metric_value:
+                    update_vals[metric_value['metric'].lower()] = metric_value['totalValue']['value']
+
+            location_report.update(update_vals)
 
     @api.multi
     def search_report(self):
         self.ensure_one()
 
-        if self.env['google.ads'].need_authorize():
-            raise UserError(_("Need to authorize with Google Ads first"))
-
-        #Get a new access token
-        refresh_token = self.env['ir.config_parameter'].get_param('google_ads_refresh_token')
-        all_token = self.env['google.service']._refresh_google_token_json(refresh_token, 'ads')
-        access_token = all_token.get('access_token')
-
-        manager_customer_id = self.env['ir.default'].get('sem.settings', 'google_ads_manager_account_customer_id').replace("-","")
-        developer_token = self.env['ir.default'].get('sem.settings', 'google_ads_developer_token')
-        headers = {'login-customer-id': manager_customer_id, 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + access_token, 'developer-token': developer_token}
-
-        # TODO allow for other languages
-        language = 'languageConstants/1000' #English
-
-        # TODO allow this to be configured
-        keyword_plan_network = "GOOGLE_SEARCH"
-
         search_report = self.env['sem.report.search'].create({'website_id': self.id})
 
-        for keyword in self.keyword_ids:
-            if keyword.active:
-                for geo_target in keyword.geo_target_ids:
-
-                    payload = json.dumps({
-                        "language": language,
-                        "geo_target_constants":
-                        [
-                            "geoTargetConstants/" + geo_target.location_id
-                        ],
-                        "keyword_plan_network": keyword_plan_network,
-                        "keyword_seed":
-                        {
-                            "keywords":
-                            [
-                                keyword.name
-                            ]
-                        }
-                    })
-
-                    response_string = requests.post("https://googleads.googleapis.com/v2/customers/" + manager_customer_id + ":generateKeywordIdeas", data=payload, headers=headers)
-
-                    response_string_json = json.loads(response_string.text)
-
-                    monthly_searches = "-"
-                    competition = "-"
-
-                    if "results" in response_string_json:
-                        exact_match = response_string_json['results'][0]
-                        if exact_match['text'] == keyword.name.lower():
-                            monthly_searches = exact_match['keywordIdeaMetrics']['avgMonthlySearches']
-                            if 'competition' in exact_match['keywordIdeaMetrics']:
-                                competition = exact_match['keywordIdeaMetrics']['competition']
-
-                    self.env['sem.report.search.result'].create({'search_id': search_report.id, 'keyword_id': keyword.id, 'geo_target_id': geo_target.id, 'monthly_searches': monthly_searches, 'competition': competition})
+        for search_context in self.search_context_ids:
+            search_insight = search_context.get_insight()
+            search_insight['search_id'] = search_report.id
+            search_insight['search_context_id'] = search_context.id
+            self.env['sem.report.search.result'].create(search_insight)
 
         return {
             'name': 'Search Report',
@@ -275,74 +241,15 @@ class SemClientWebsite(models.Model):
         }
 
     @api.multi
-    def maps_report(self):
+    def add_search_context(self):
         self.ensure_one()
 
-        for keyword in self.keyword_ids:
-            if keyword.active:
-                for geo_target in keyword.geo_target_ids:
-                    zoom_level = str(geo_target.map_zoom_level)
-                    url = "https://www.google.com/maps/search/" + keyword.name + "/@" + geo_target.latitude + "," + geo_target.lonhitude + "," + zoom_level + "z"
-
         return {
-            'name': 'Maps Report',
+            'name': 'Create Search Context',
             'view_type': 'form',
             'view_mode': 'form',
-            'res_model': 'sem.report.map',
-            'res_id': maps_report.id,
-            'type': 'ir.actions.act_window'
-        }
-
-    @api.multi
-    def ranking_report(self):
-        self.ensure_one()
-
-        ranking_report = self.env['sem.report.ranking'].create({'website_id': self.id})
-
-        for keyword in self.keyword_ids:
-            if keyword.active:
-                for geo_target in keyword.geo_target_ids:
-
-                    create_dict = {'ranking_id': ranking_report.id, 'keyword_id': keyword.id, 'geo_target_id': geo_target.id}
-
-                    # For Google this returns search_url, for Bing this returns rank, link_url and item_ids
-                    create_dict.update(geo_target.get_ranking(self.url, keyword))
-
-                    # Create the ranking entry even if the client's website is not found
-                    self.env['sem.report.ranking.result'].create(create_dict)
-
-        return {
-            'name': 'Ranking Report',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'sem.report.ranking',
-            'res_id': ranking_report.id,
-            'type': 'ir.actions.act_window'
-        }
-
-class SemClientWebsiteKeyword(models.Model):
-
-    _name = "sem.client.website.keyword"
-
-    website_id = fields.Many2one('sem.client.website', string="Website")
-    name = fields.Char(string="Keyword")
-    active = fields.Boolean(string="Active", default="True", help="Keywords can not be deleted only archived so new reports don't show them")
-    geo_target_ids = fields.Many2many('sem.geo_target', string="Geo Targets")
-
-    @api.multi
-    def find_geo_targets(self):
-        self.ensure_one()
-
-        # Clean up list from previous searches
-        for geo_result in self.env['sem.geo_target.wizard.record'].search([]):
-            geo_result.unlink()
-
-        return {
-            'name': 'Find Geo Targets',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'sem.geo_target.wizard',
+            'res_model': 'sem.search.context.wizard',
             'target': 'new',
-            'context': {'default_keyword_id': self.id},
+            'context': {'default_type': 'search', 'default_website_id': self.id},
             'type': 'ir.actions.act_window'
-        } 
+        }