+++ sem_seo_toolkit/models/sem_settings.py
@@ -32,54 +32,11 @@ class SEMSettings(models.Model):
     google_ads_client_id = fields.Char(string="Google Ads Client ID")
     google_ads_client_secret = fields.Char(string="Google Ads Client Secret")
     google_ads_manager_account_customer_id = fields.Char(string="Google Ads Manager Account Customer ID")
-    #google_my_business_client_id = fields.Char(string="Google My Business Client ID")
-    #google_my_business_client_secret = fields.Char(string="Google My Business Client Secret")
+    google_my_business_client_id = fields.Char(string="Google My Business Client ID")
+    google_my_business_client_secret = fields.Char(string="Google My Business Client Secret")
     bing_web_search_api_key = fields.Char(string="Bing Web Search v7 API Key")
-    bing_maps_api_key = fields.Char(string="Bing maps API Key")
+    bing_maps_api_key = fields.Char(string="Bing Maps API Key")
     tracking_code = fields.Text(string="Tracking Code", default=_default_tracking_code)
-
-    def test_google_my_business(self):
-         # Have to wait for access before I can start feature...
-        response = requests.get("https://mybusiness.googleapis.com/v4/accounts")
-        _logger.error(response.text)
-
-        #All metrics within last 30 days
-        payload = json.dumps({
-            "locationNames": [
-                "accounts/account_name/locations/locationId"
-            ],
-            "basicRequest": {
-                "metricRequests": [
-                    {
-                        "metric": "ALL"
-                    }
-                ],
-                "timeRange": {
-                    "startTime": "2016-10-12T01:01:23.045123456Z",
-                    "endTime": "2017-01-10T23:59:59.045123456Z"
-                }
-            }
-        })
-
-        response = requests.post("https://mybusiness.googleapis.com/v4/{name=accounts/*}/locations:reportInsights", data=payload)
-        _logger.error(response.text)
-        response_json = json.loads(response.text)
-        create_vals = {}
-        create_vals['queries_direct'] = response_json['QUERIES_DIRECT']
-        create_vals['queries_indirect'] = response_json['QUERIES_INDIRECT']
-        create_vals['queries_chain'] = response_json['QUERIES_CHAIN']
-        create_vals['views_maps'] = response_json['VIEWS_MAPS']
-        create_vals['views_search'] = response_json['VIEWS_SEARCH']
-        create_vals['actions_website'] = response_json['ACTIONS_WEBSITE']
-        create_vals['actions_phone'] = response_json['ACTIONS_PHONE']
-        create_vals['actions_driving_directions'] = response_json['ACTIONS_DRIVING_DIRECTIONS']
-        create_vals['photos_views_merchant'] = response_json['PHOTOS_VIEWS_MERCHANT']
-        create_vals['photos_views_customers'] = response_json['PHOTOS_VIEWS_CUSTOMERS']
-        create_vals['photos_count_merchant'] = response_json['PHOTOS_COUNT_MERCHANT']
-        create_vals['photos_count_customers'] = response_json['PHOTOS_COUNT_CUSTOMERS']
-        create_vals['local_post_views_search'] = response_json['LOCAL_POST_VIEWS_SEARCH']
-        create_vals['local_post_actions_call_to_action'] = response_json['LOCAL_POST_ACTIONS_CALL_TO_ACTION']
-        self.env['sem.report.google.business'].create(create_vals)
 
     @api.multi
     def google_ads_authorize(self):
@@ -87,6 +44,14 @@ class SEMSettings(models.Model):
 
         return_url = request.httprequest.host_url + "web"
         url = self.env['google.service']._get_authorize_uri(return_url, 'ads', scope='https://www.googleapis.com/auth/adwords')
+        return {'type': 'ir.actions.act_url', 'url': url, 'target': 'self'}
+
+    @api.multi
+    def google_my_business_authorize(self):
+        self.ensure_one()
+
+        return_url = request.httprequest.host_url + "web"
+        url = self.env['google.service']._get_authorize_uri(return_url, 'business', scope='https://www.googleapis.com/auth/business.manage')
         return {'type': 'ir.actions.act_url', 'url': url, 'target': 'self'}
 
     @api.multi
@@ -100,6 +65,8 @@ class SEMSettings(models.Model):
         self.env['ir.default'].set('sem.settings', 'google_ads_manager_account_customer_id', self.google_ads_manager_account_customer_id)
         self.env['ir.default'].set('sem.settings', 'bing_web_search_api_key', self.bing_web_search_api_key)
         self.env['ir.default'].set('sem.settings', 'bing_maps_api_key', self.bing_maps_api_key)
+        self.env['ir.config_parameter'].set_param('google_business_client_id', self.google_my_business_client_id)
+        self.env['ir.config_parameter'].set_param('google_business_client_secret', self.google_my_business_client_secret)
 
     @api.model
     def get_values(self):
@@ -112,6 +79,8 @@ class SEMSettings(models.Model):
             google_ads_client_secret=self.env['ir.config_parameter'].get_param('google_ads_client_secret'),
             google_ads_manager_account_customer_id=self.env['ir.default'].get('sem.settings', 'google_ads_manager_account_customer_id'),
             bing_web_search_api_key=self.env['ir.default'].get('sem.settings', 'bing_web_search_api_key'),
-            bing_maps_api_key=self.env['ir.default'].get('sem.settings', 'bing_maps_api_key')
+            bing_maps_api_key=self.env['ir.default'].get('sem.settings', 'bing_maps_api_key'),
+            google_my_business_client_id=self.env['ir.config_parameter'].get_param('google_business_client_id'),
+            google_my_business_client_secret=self.env['ir.config_parameter'].get_param('google_business_client_secret')
         )
         return res